% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{Helper function untuk membuat diagnostic plots}
\usage{
diagnostic_plots(model)
}
\arguments{
\item{model}{Objek model lm}
}
\description{
Helper function untuk membuat diagnostic plots
}
