% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_report}
\alias{generate_report}
\title{Generate laporan uji asumsi klasik}
\usage{
generate_report(hasil_uji, file = NULL)
}
\arguments{
\item{hasil_uji}{Hasil dari uji_asumsi_klasik}

\item{file}{Nama file output (opsional)}
}
\description{
Generate laporan uji asumsi klasik
}
