% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_info}
\alias{model_info}
\title{Extract model information}
\usage{
model_info(model)
}
\arguments{
\item{model}{Objek model lm}
}
\value{
List berisi informasi model
}
\description{
Extract model information
}
