% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perbaikan_heteroskedastisitas.R
\name{plot.perbaikan_heteroskedastisitas}
\alias{plot.perbaikan_heteroskedastisitas}
\title{Plot diagnostics untuk perbaikan heteroskedastisitas}
\usage{
\method{plot}{perbaikan_heteroskedastisitas}(hasil, ...)
}
\arguments{
\item{hasil}{Hasil dari perbaikan_heteroskedastisitas()}
}
\description{
Plot diagnostics untuk perbaikan heteroskedastisitas
}
