% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perbaikan_heteroskedastisitas.R
\name{print.perbaikan_heteroskedastisitas}
\alias{print.perbaikan_heteroskedastisitas}
\title{Print method untuk perbaikan_heteroskedastisitas}
\usage{
\method{print}{perbaikan_heteroskedastisitas}(x, ...)
}
\description{
Print method untuk perbaikan_heteroskedastisitas
}
