% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perbaikan_multikolinearitas.R
\name{print.perbaikan_multikolinearitas}
\alias{print.perbaikan_multikolinearitas}
\title{Print method untuk perbaikan_multikolinearitas}
\usage{
\method{print}{perbaikan_multikolinearitas}(x, ...)
}
\description{
Print method untuk perbaikan_multikolinearitas
}
