% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{print.transformasi_hasil}
\alias{print.transformasi_hasil}
\title{Print method untuk transformasi_hasil}
\usage{
\method{print}{transformasi_hasil}(x, ...)
}
\description{
Print method untuk transformasi_hasil
}
