% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uji_asumsi_klasik.R
\name{print.uji_asumsi_klasik_hasil}
\alias{print.uji_asumsi_klasik_hasil}
\title{Print method untuk uji_asumsi_klasik_hasil}
\usage{
\method{print}{uji_asumsi_klasik_hasil}(x, ...)
}
\description{
Print method untuk uji_asumsi_klasik_hasil
}
