% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uji_asumsi_klasik_updated.R
\name{print.uji_asumsi_klasik_plus}
\alias{print.uji_asumsi_klasik_plus}
\title{Print method untuk uji_asumsi_klasik_plus}
\usage{
\method{print}{uji_asumsi_klasik_plus}(x, ...)
}
\description{
Print method untuk uji_asumsi_klasik_plus
}
