% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uji_homoskedastisitas.R
\name{print.uji_homoskedastisitas}
\alias{print.uji_homoskedastisitas}
\title{Print method untuk uji_homoskedastisitas}
\usage{
\method{print}{uji_homoskedastisitas}(x, ...)
}
\description{
Print method untuk uji_homoskedastisitas
}
