% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uji_multikolinearitas.R
\name{print.uji_multikolinearitas}
\alias{print.uji_multikolinearitas}
\title{Print method untuk uji_multikolinearitas}
\usage{
\method{print}{uji_multikolinearitas}(x, ...)
}
\description{
Print method untuk uji_multikolinearitas
}
