% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi.R
\name{transformasi_otomatis}
\alias{transformasi_otomatis}
\title{Transformasi Variabel}
\usage{
transformasi_otomatis(
  data,
  formula,
  max_iter = 3,
  alpha = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data frame}

\item{formula}{Formula model}

\item{max_iter}{Maksimum iterasi transformasi (default = 3)}

\item{alpha}{Tingkat signifikansi (default = 0.05)}

\item{verbose}{Logical, apakah menampilkan proses (default = TRUE)}
}
\value{
List berisi model terbaik dan hasil transformasi
}
\description{
Melakukan transformasi pada variabel dependen untuk mengatasi pelanggaran asumsi
}
