% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uji_asumsi_klasik.R
\name{uji_asumsi_klasik}
\alias{uji_asumsi_klasik}
\title{Uji Asumsi Klasik Lengkap}
\usage{
uji_asumsi_klasik(model, alpha = 0.05, plot = TRUE, transform = FALSE)
}
\arguments{
\item{model}{Objek model lm}

\item{alpha}{Tingkat signifikansi (default = 0.05)}

\item{plot}{Logical, apakah menampilkan plot (default = TRUE)}

\item{transform}{Logical, apakah melakukan transformasi otomatis jika asumsi dilanggar (default = FALSE)}
}
\value{
List berisi hasil semua uji asumsi klasik
}
\description{
Melakukan semua uji asumsi klasik sekaligus
}
