% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uji_autokorelasi.R
\name{uji_autokorelasi}
\alias{uji_autokorelasi}
\title{Uji Autokorelasi}
\usage{
uji_autokorelasi(model, alpha = 0.05, plot = TRUE)
}
\arguments{
\item{model}{Objek model lm}

\item{alpha}{Tingkat signifikansi (default = 0.05)}

\item{plot}{Logical, apakah menampilkan plot (default = TRUE)}
}
\value{
List berisi hasil uji autokorelasi
}
\description{
Melakukan berbagai uji autokorelasi pada model regresi
}
