% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uji_homoskedastisitas.R
\name{uji_homoskedastisitas}
\alias{uji_homoskedastisitas}
\title{Uji Homoskedastisitas}
\usage{
uji_homoskedastisitas(model, alpha = 0.05, plot = TRUE)
}
\arguments{
\item{model}{Objek model lm}

\item{alpha}{Tingkat signifikansi (default = 0.05)}

\item{plot}{Logical, apakah menampilkan plot (default = TRUE)}
}
\value{
List berisi hasil uji homoskedastisitas
}
\description{
Melakukan berbagai uji homoskedastisitas pada model regresi
}
