% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uji_multikolinearitas.R
\name{uji_multikolinearitas}
\alias{uji_multikolinearitas}
\title{Uji Multikolinearitas}
\usage{
uji_multikolinearitas(model, threshold_vif = 10, plot = TRUE)
}
\arguments{
\item{model}{Objek model lm}

\item{threshold_vif}{Nilai threshold untuk VIF (default = 10)}

\item{plot}{Logical, apakah menampilkan plot (default = TRUE)}
}
\value{
List berisi hasil uji multikolinearitas
}
\description{
Melakukan uji multikolinearitas pada model regresi
}
