% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uji_normalitas.R
\name{uji_normalitas}
\alias{uji_normalitas}
\title{Uji Normalitas Residual}
\usage{
uji_normalitas(model, alpha = 0.05, plot = TRUE)
}
\arguments{
\item{model}{Objek model lm}

\item{alpha}{Tingkat signifikansi (default = 0.05)}

\item{plot}{Logical, apakah menampilkan plot (default = TRUE)}
}
\value{
List berisi hasil uji normalitas
}
\description{
Melakukan berbagai uji normalitas pada residual model regresi
}
\examples{
model <- lm(mpg ~ wt + hp, data = mtcars)
uji_normalitas(model)
}
